Populate Street Dropdown With API Data
This commit changed the structure of the form to use the API data to populate the dropdowns.  Unfortunately, the data model for Neighborhood only has one item in it, so I'm iterating over the Street objects instead of doing a fresh call for the Neighborhood objects.  Will have to fix that in the API in the future.  Main concern now is making sure the selection of a street returns the correct data when the checkboxes (even / odd) are checked.