Create Stateful Checkboxes
This commit creates the checkboxes needed to display the street details on click.  In creating them, however, it became clear that the origanization of this project could be re-worked for a cleaner App.js file.  It also would be helpful to break up that file into re-usable components now that I have the main pieces laid out.  I also had to think through which parts actually need state.  Next steps are to create the components.