Create StreetSelectorContainer And Move Components Into It
This commit created the container for the neighborhood and street dropdowns as well as the even/odd checkboxes.  Unfortunately, there is a problem iterating over `this.selectedCheckboxes` in App.js, because it lives in StreetSelectorContainer.js.  It is unclear how to resolve this as it is not recommended to try to pass the child elements up to the parent via a function...  Hmm...