Create All Containers And Begin Linking JS Events With Elements
This commit created all of the remaining containers needed.  It also began linking the loading of the street objects with the selection of the correct neighborhood.  Currently this is quite clunky because the API only has one model.

The next step is to add the Neighborhood and BufferTime models to the API.